+++ src/components/ParkView/Components/VirtualGarden/Container.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useMemo, useState } from 'react';
 
 import { useParkViewContext } from '../../Context/ParkViewProvider';
+import { useGardenApi } from '../../hooks/useGardenApi';
 import { DataSet } from '../../Models/data';
 import { GardenApi } from '../../Models/gardenApi';
 import { PostGroupBySorting, PreGroupByFiltering } from '../../Models/gardenOptions';
@@ -94,40 +95,3 @@ export const VirtualContainer = <T extends unknown>({
         </Container>
     );
 };
-
-interface GardenApiConstructor {
-    createGardenApi: () => GardenApi;
-}
-
-export function useGardenApi(): GardenApiConstructor {
-    const {
-        groupByKeys,
-        customGroupByKeys,
-        setGroupKeys,
-        setCustomGroupKeys,
-        setGardenKey,
-        setCustomState,
-        customState,
-        gardenKey,
-    } = useParkViewContext();
-
-    function createGardenApi(): GardenApi {
-        return {
-            mutations: {
-                setCustomGroupKeys: setCustomGroupKeys,
-                setCustomState: setCustomState,
-                setGardenKey: setGardenKey,
-                setGroupKeys: setGroupKeys,
-            },
-            states: {
-                getCurrentGroupByKeys: () => groupByKeys as string[],
-                getCustomGroupByKeys: () => customGroupByKeys,
-                getCustomState: () => customState,
-                getGardenKey: () => gardenKey,
-            },
-        };
-    }
-    return {
-        createGardenApi: createGardenApi,
-    };
-}
